Add files via upload
Create a Function to retrieve the number of years of service for a specific employee.  The function should accept the employee ID as a parameter and return the number of years of service. Add error handling to account for an invalid employee ID.

Create a stored procedure called UPD_SAL to update the  salaries
for a specific job in the EMP table.